% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allDefaultModels}
\alias{allDefaultModels}
\title{Returns all default models}
\usage{
allDefaultModels()
}
\value{
A list of all the default model functions.
}
\description{
This is a convenience function that returns a list of all the default
models.
}

