% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assignColumnNames.R
\name{assignColumnNames}
\alias{assignColumnNames}
\title{Assign Column Names}
\usage{
assignColumnNames(columnNames, data, envir)
}
\arguments{
\item{columnNames}{A named character vector describing which variables
(the names) map to which columns of data (the values).}

\item{data}{The dataset being analysed.  This is used to ensure that all
elements of columnNames are in fact in data.}

\item{env}{The environment where the column name variables should be
assigned.  Typically, this will be the environment of the function which
called this function.  See examples.}
}
\value{
No value is returned.  However, values are assigned in the
environment which called this function.
}
\description{
This function takes the values of columnNames and assigns them to the
variables defined by names(columnNames), respectively.  This assignment is
done in the environment passed to this function by env.  Not all variables
value/name pairs are assigned, however; only the variables needed by the
functions of this package (execute defaultColumnNames() to see this list).
}
\details{
This function is used in combination with the columnNames argument to load
variables into the environment of a particular function.  That function can
then access all these variables (which describe the columns of data)
without creating global variables.
}
\examples{
f = function(){
    assignColumnNames( defaultColumnNames, okrapd, environment() )
    print(productionValue)
}
#productionValue gets assigned a character value from defaultColumnNames:
f()
#However, the assignment is only within f().  productionValue doesn't exist
#in the global environment:
exists("productionValue")
}

