% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assignParameters.R
\name{assignParameters}
\alias{assignParameters}
\title{Assign Parameters}
\usage{
assignParameters(parameterList, environment = .GlobalEnv)
}
\arguments{
\item{parameterList}{A named character vector with the variables that
should be assigned.}

\item{environment}{The environment where the column name variables should be
assigned.  Defaults to the global environment.}
}
\value{
No value is returned.  However, values are assigned to the passed
environment.  Since values are typically assigned in the global environment,
lockBinding is also called on these variables to ensure they are not easily
changed.  If the variables need to be changed, unlockBinding can be called
before reassigning.  Alternatively, one may use reassignGlobalVariable.
}
\description{
This function takes the values of a parameter list and assigns them to the
variables defined by names(parameterList), see defaultImputationParameters
and defaultProcessingParameters for examples.  This assignment is done in
the environment passed to this function by env.  All value/name pairs are
assigned, as a check is made to ensure the parameters in parameterList are
exactly the ones needed for the processing or imputation.
}
\details{
This function is used in combination with the parameterList argument to load
variables into a particular environment.  Functions can then access all
these variables (which describe the columns of the dataset) without needing
to pass long lists through.
}

