% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/balanceAreaHarvested.R
\name{balanceAreaHarvested}
\alias{balanceAreaHarvested}
\title{Function to compute area harvested when new production and yield
are given.}
\usage{
balanceAreaHarvested(columnNames, newMethodFlag, flagTable = faoswsFlagTable,
  data)
}
\arguments{
\item{columnNames}{See same argument at ?imputeProductionDomain.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{data}{The data.table object containing the data.}
}
\description{
Function to compute area harvested when new production and yield
are given.
}

