% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class-ensembleModel.R
\name{checkEnsembleModel}
\alias{checkEnsembleModel}
\title{Class for Ensemble Model}
\usage{
checkEnsembleModel(object)
}
\arguments{
\item{model}{The function defining how the model is fit to the data.  The
function should take arguments data, value, flag, columnNames, and
missingIndex (see ?extendSimpleModel).}

\item{extrapolationRange}{How many time steps outside of the data is this
model valid for?  Should be a positive integer (or Inf).}

\item{level}{The level at which this model is applied.  Currently, must be
one of "countryCommodity" or "commodity".}
}
\description{
Ensemble models contain a definition (function) of the algorithm used to
fit the model as well as additional parameters:\n
Extrapolation range: defines how far outside the range of the data this
model is valid.\n
Level: defines if this model operates on each country-commodity pair
individually ("countryCommodity" level) or all countries for a fixed
commodity ("commodity" level) at once.
}

