% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeEnsemble.R
\name{computeEnsemble}
\alias{computeEnsemble}
\title{Function to combine the ensembles}
\usage{
computeEnsemble(fits, weights)
}
\arguments{
\item{fits}{A list of fitted values from models.}

\item{weights}{A vector for the weights of each model.}
}
\description{
Function to combine the ensembles
}

