% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeEnsembleWeight}
\alias{computeEnsembleWeight}
\title{Function to compute the weights of the ensemble models}
\usage{
computeEnsembleWeight(x, fits, restrictWeights = TRUE, maximumWeights = 0.7,
  errorType = "mse", ensembleModel = NULL)
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{fits}{The fitted value from the models.}

\item{restrictWeights}{Whether a maximum weight restriction should
be imposed.}

\item{maximumWeights}{The maximum weight to be imposed, must be
between [0.5, 1].}

\item{errorType}{Specifies what type of error model should be used to
compute model weights.  Currently, one of "mse" (mean squared error
between the observed values and the model values) or "loocv" (error
computed via leave-one-out cross-validation).}

\item{ensembleModel}{A list of the models fit.  This is only needed if
errorType = "loocv".}
}
\description{
Function to compute the weights of the ensemble models
}

