% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeLoocvFits.R
\name{computeLoocvFits}
\alias{computeLoocvFits}
\title{Function to compute the fitted values from Leave One Out Cross Validation}
\usage{
computeLoocvFits(data, cvGroup, imputationParameters)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{cvGroup}{A vector of the same length as nrow(data).  Entries of the
vector should be integers from 1 to the number of cross-validation groups
(typically 10).  This should be randomly assigned, and is usually created
by ensembleImpute.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
A list of the same length as ensembleModels which contains the
fitted values corresponding to the loocv procedure.  The ith element of the
list corresponds to the fitting of the ith element of ensembleModels.
}
\description{
Function to compute the fitted values from Leave One Out Cross Validation
}

