% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/containInfo.R
\name{containInfo}
\alias{containInfo}
\title{Function to determine whether the data contains any information.}
\usage{
containInfo(value, flag, processingParameters)
}
\arguments{
\item{value}{A numeric vector to be checked}

\item{flag}{The character vector representing the observation flag
corresponding to the value.}

\item{naFlag}{The observation flag which corresponds to missing
value.}
}
\description{
If the data contains only a missing value or zero then it is
marked as no information.
}

