% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultArima.R
\name{defaultArima}
\alias{defaultArima}
\title{The default ARIMA model for the ensemble model.}
\usage{
defaultArima(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default ARIMA model for the ensemble model.
}

