% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultColumnNames.R
\name{defaultColumnNames}
\alias{defaultColumnNames}
\title{Default Column Names}
\usage{
defaultColumnNames()
}
\value{
A named vector, typically to be passed to another function.  It
may be necessary to modify some values of this vector so that they are
appropriate for your particular dataset.
}
\description{
This function generates a named vector of column names.  It's useful as a
starting point for the columnNames argument required by many functions in
this package.
}

