% Generated by roxygen2 (4.0.1): do not edit by hand
\name{defaultExp}
\alias{defaultExp}
\title{The default exponential model for the ensemble model.}
\usage{
defaultExp(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default exponential model for the ensemble model.
}

