% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultImputationParameters.R
\name{defaultImputationParameters}
\alias{defaultImputationParameters}
\title{Default Imputation Parameters}
\usage{
defaultImputationParameters()
}
\value{
Returns a list of the default parameters used in the ensemble
imputation algorithms.
}
\description{
This function can be used to generate the input parameters for the
ensemble imputation code.  This is a good way to get a list of the required
parameters and then modify parameters to match your particular
configuration.
}

