% Generated by roxygen2 (4.0.1): do not edit by hand
\name{defaultLm}
\alias{defaultLm}
\title{The default linear regression model for the ensemble model.}
\usage{
defaultLm(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default linear regression model for the ensemble model.
}

