% Generated by roxygen2 (4.0.1): do not edit by hand
\name{defaultLoess}
\alias{defaultLoess}
\title{The default exponential model for the ensemble model.}
\usage{
defaultLoess(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default exponential model for the ensemble model.
}

