% Generated by roxygen2 (4.0.1): do not edit by hand
\name{defaultLogistic}
\alias{defaultLogistic}
\title{The default logistic model for the ensemble model.}
\usage{
defaultLogistic(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default logistic model for the ensemble model.
}

