% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultMean.R
\name{defaultMean}
\alias{defaultMean}
\title{The default mean model for the ensemble model.}
\usage{
defaultMean(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default mean model for the ensemble model.
}

