% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultMixedModel.R
\name{defaultMixedModel}
\alias{defaultMixedModel}
\title{Mixed Model for Imputation}
\usage{
defaultMixedModel(data, value, flag, columnNames, maxdf = 5, weights = NULL,
  modelFormula)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{value}{Which variable should be imputed?  This should be one of the
column names of data corresponding to productionValue, yieldValue, or
areaHarvestedValue.}

\item{flag}{Column name (of data) of the flag variable corresponding to the value
argument, i.e. "productionObservationFlag", "yieldObservationFlag", or
"areaHarvestedObservationFlag".}

\item{columnNames}{See the same argument at ?imputeProductionDomain.}

\item{maxdf}{The maximum degree of freedom for the spline.}

\item{weights}{The weights for the observation}

\item{modelFormula}{Formula specifying how the dependent variable (value)
depends on the other columns of data.  Should be a valid mixed model
formula, as it will be passed to lmer (R's mixed model function).}
}
\value{
Returns a vector of the estimated/imputed values.  If a value
existed in the original data, then an NA is returned in that location.
}
\description{
This function imputes missing values through linear mixed model.
}

