% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultProcessingParameters.R
\name{defaultProcessingParameters}
\alias{defaultProcessingParameters}
\title{Default Processing Parameters}
\usage{
defaultProcessingParameters()
}
\value{
Returns a list of the default parameters used in the data
pre-processing algorithm.
}
\description{
This function can be used to generate the input parameters for the
data pre-processing code.  This is a good way to get a list of the required
parameters and then modify parameters to match your particular
configuration.
}

