% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensureColumnNames.R
\name{ensureColumnNames}
\alias{ensureColumnNames}
\title{Ensure Column Names}
\usage{
ensureColumnNames(columnNames, data)
}
\arguments{
\item{columnNames}{A named character vector describing which variables
(the names) map to which columns of data (the values).}

\item{data}{The dataset being analysed.  This is used to ensure that all
elements of columnNames are in fact in data.}
}
\value{
No value is returned.  However, an error will be raised if the
columnNames vector is invalid.
}
\description{
In addition, this function ensures that the columnNames argument is valid.
In particular, that means that all elements of columnNames are valid column
names in data.  Additionally, this function also ensures that all of the
necessary columns are provided (necessary for other functions in this
package).
}

