% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensureImputationData.R
\name{ensureImputationData}
\alias{ensureImputationData}
\title{Ensure Imputation Data}
\usage{
ensureImputationData(data)
}
\arguments{
\item{data}{A data.table containing the data.}
}
\description{
This function is designed to ensure that the provided dataset is valid.  In
particular, it coerces column types: all values are coerced to numeric (
instead of integer, which can cause problems) and all flags are coerced to
character (instead of logical, which occurs if the flag is set to NA).
Also, it ensures data is a data.table.
}
\details{
Names for the columns of this data.table should exist in the calling
environment (and likely in the global environment).  For example, a
character variable called productionValue should specify the column name of
data which corresponds to the production value.  Other variables should
also exist, such as parametersAssigned and ensuredData.  These will all
typically be assigned by calling the assignParameters function with an
argument of either defaultImputationParameters().
}

