% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensureImputationParameters.R
\name{ensureImputationParameters}
\alias{ensureImputationParameters}
\title{Ensure Imputation Parameters are Valid}
\usage{
ensureImputationParameters(parameterList)
}
\arguments{
\item{parameterList}{A list of the imputation parameters, such as one
generated by defaultImputationParameters().}
}
\value{
No value is returned, but errors are thrown if the input parameters
are not as expected.
}
\description{
Ensure Imputation Parameters are Valid
}

