% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensureProcessingParameters.R
\name{ensureProcessingParameters}
\alias{ensureProcessingParameters}
\title{Ensure Processing Parameters are Valid}
\usage{
ensureProcessingParameters(parameterList)
}
\arguments{
\item{parameterList}{A list of the processing parameters, such as one
generated by defaultProcessingParameters().}
}
\value{
No value is returned, but errors are thrown if the input parameters
are not as expected.
}
\description{
Ensure Processing Parameters are Valid
}

