% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extendSimpleModel.R
\name{extendSimpleModel}
\alias{extendSimpleModel}
\title{Extend Simple Model}
\usage{
extendSimpleModel(data, model, value, flag, columnNames)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{model}{The model to be applied to each individual time series.
Typically, this will be a function such as one from allDefaultModels().}

\item{value}{Which variable should be imputed?  This should be one of the
column names of data, typically corresponding to productionValue,
yieldValue, or areaHarvestedValue.}

\item{flag}{Column name (of data) of the flag variable corresponding to the
value argument, i.e. "productionObservationFlag", "yieldObservationFlag",
or "areaHarvestedObservationFlag".}

\item{columnNames}{See the same argument at ?imputeProductionDomain.}
}
\value{
Returns a vector of the estimated/imputed values.  If a value
existed in the original data, then an NA is returned in that location.
}
\description{
This function takes any model which is designed to run on a simple time
series (i.e. at the commodity-country level) and applies it to each country
in turn, thus making it a commodity level model.
}

