% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faoswsProductionImputation-package.R
\docType{package}
\name{faoswsProductionImputation-package}
\alias{faoswsProductionImputation-package}
\title{Package to perform imputation for the FAOSTAT production
domain.}
\description{
This package provides the imputation methodology designed for the
FAO FAOSTAT production domain.
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}

Joshua M. Browning \email{joshua.browning@fao.org}
}
\keyword{package}

