% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getInverseWeights.R
\name{getInverseWeights}
\alias{getInverseWeights}
\title{Get Ensemble Weights via Inverse Errors}
\usage{
getInverseWeights(data, fits, imputationParameters)
}
\arguments{
\item{data}{The data object containing the observations to impute.}

\item{fits}{A list of the fitted values.  These may be estimated via leave
one out cross-validation or directly.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
A data.table containing the weight for each model within each byKey
group, as well as a few other (currently unused) statistics.
}
\description{
Weights used to construct the final ensemble from the individual models are
computed by comparing the errors.  For each model, the errors are computed
(simply comparing observed values in data to fitted values in fits) and
then the inverse (i.e. 1/) these errors are used as the weights (after
rescaling to 1).  For example, if three models have errors of 1, 2, and 3
then the inverse errors are 1, 1/2, and 1/3.  We rescale these weights to
add to 1 and get 6/11, 3/11, and 2/11.
}

