% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getObservedExtrapolationRange.R
\name{getObservedExtrapolationRange}
\alias{getObservedExtrapolationRange}
\title{Get Observed Extrapolation Range}
\usage{
getObservedExtrapolationRange(x)
}
\arguments{
\item{x}{The vector of data.}
}
\value{
A vector of the same length as x.
}
\description{
This function takes a vector of data (with posisbly missing values) and
returns a vector of the same length.  If the corresponding value of x is
not missing or is an interpolation, the returned value is 0.  If the
corresponding value of x is missing and is an extrapolation, then the
returned value is the "distance" (in terms of number of observations) to
the closest non-missing value.  See examples.
}
\examples{
x = c(NA, NA, NA, 1:7, NA, NA)
getObservedExtrapolationRange(x)
}

