% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getWeightMatrix.R
\name{getWeightMatrix}
\alias{getWeightMatrix}
\title{Get Weight Matrix}
\usage{
getWeightMatrix(data, value, byKey, yearValue, w, ensembleModels)
}
\arguments{
\item{data}{The data.table containing the data.}

\item{value}{The column name of data which contains the variable being
imputed.}

\item{byKey}{The column name of data which specifies the grouping
(typically the areaCode).}

\item{yearValue}{The column name of data which specifies the year.}

\item{w}{The weights data.table, typically as produced in
computeEnsembleWeight.  There should be at least three columns: byKey,
model, and weight.  Weight gives the model weight for model within the
byKey group.}

\item{ensembleModels}{The list of ensemble models used.  Each element of
this list should be of class ensembleModel.  This is used to determine
the extrapolation range of each model in the ensemble.}
}
\value{
A matrix of weights that can be multiplied by the fitted models to
give the imputed values.  Rows corresponding to non-missing values in data
have values of NA.
}
\description{
Initially, a weight is computed for each model and byKey.  However, some
models are not valid for some observations (as certain models are limited
in how far they can extrapolate outside the range of the data).  Thus, the
final weight for each ensemble model at each observation will depend on
that models performance for that byKey group as well as if that model is
valid at that point.
}
\details{
This function creates a weight matrix to use in constructing the final
ensemble.  If F is a nxk matrix (n = number of observations, k = number of
models) containing the fitted models, then this function constructs W,
another nxk matrix of weights.  The final ensemble estimate for observation
i can be computed by sum( F[i,]*W[i,] ).
}

