% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeProduction.R
\name{imputeProduction}
\alias{imputeProduction}
\title{Function to impute production}
\usage{
imputeProduction(columnNames, imputationFlag = "I", newMethodFlag, data,
  restrictWeights = TRUE, maximumWeights = 0.7,
  ensembleModel = allDefaultModels(),
  modelExtrapolationRange = getDefaultRange(ensembleModel),
  flagTable = faoswsFlagTable, errorType = "loocv",
  errorFunction = function(x) mean(x^2))
}
\arguments{
\item{columnNames}{See the same argument at ?imputeProductionDomain}

\item{imputationFlag}{Flag value for new imputation values.}

\item{data}{The data.table object containing the data.}

\item{restrictWeights}{Whether a maximum weight restriction should
be imposed.}

\item{maximumWeights}{The maximum weight to be imposed, must be
between [0.5, 1].}

\item{ensembleModel}{A list of models to be used to build the
ensemble.}

\item{modelExtrapolationRange}{See ?computeEnsembleWeight.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{errorType}{See ?computeErrorRate.}

\item{errorFunction}{See ?computeEnsembleWeight.}
}
\description{
This is a wrapper of the ensemble imputation for the production
domain.
}

