% Generated by roxygen2 (4.0.1): do not edit by hand
\name{imputeYield}
\alias{imputeYield}
\title{Function to impute yield}
\usage{
imputeYield(yieldValue, yieldObservationFlag, yieldMethodFlag, yearValue,
  imputationFlag, newMethodFlag, maxdf = 5, flagTable = faoswsFlagTable,
  data, weights = NULL, byKey, yieldFormula)
}
\arguments{
\item{yieldValue}{The column name of the yield variable.}

\item{yieldObservationFlag}{The observation flag of yield.}

\item{yieldMethodFlag}{The method flag of yield.}

\item{yearValue}{The column name corresponding to year.}

\item{imputationFlag}{Flag value for new imputation values.}

\item{newMethodFlag}{The new method flag to be assigned to imputed
value.}

\item{maxdf}{The maximum degree of freedom for the spline.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{data}{The data}

\item{weights}{The weights for the observation}

\item{byKey}{The unique key identifier.}
}
\description{
This function imputes the yield through linear mixed model
}

