% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loessFixedEquivalentNumberParams.R
\name{loessFixedEquivalentNumberParams}
\alias{loessFixedEquivalentNumberParams}
\title{LOESS model with fixed Equivalent Number of Parameters}
\usage{
loessFixedEquivalentNumberParams(x, enp.target = sum(!is.na(x))/2)
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{enp.target}{See ?loess.  This value specifies the equivalent number
of parameters in the loess fit, and it can be specified instead of span.}
}
\value{
A numeric vector of the same length as x, but with the model
estimates instead of the original data.
}
\description{
In some cases, the model given by defaultLoess() will fail to fit the data
well and will issue warnings.  In this case, the problem is that too much
flexibility is allowed in the loess model.  This function reduces that
flexibility by restricting the equivalent number of parameters, and thus
leads to a more stable fit.
}

