% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeCvGroup.R
\name{makeCvGroup}
\alias{makeCvGroup}
\title{Make Cross-Validation Groups}
\usage{
makeCvGroup(data, imputationParameters)
}
\arguments{
\item{data}{A data.table object containing the data.}

\item{value}{The column name of data corresponding to the value to be
imputed.}

\item{byKey}{The column name of data corresponding to the grouping
variable, typically "areaCode".}

\item{groupCount}{How many cross-validation groups should be used?}
}
\value{
A numeric vector taking values in 1:groupCount, or NA if the
corresponding observation is missing.  All groups are represented within
each individual timeseries (defined by byKey) when possible.
}
\description{
Creates a vector of cross-validation groups to be used for leave-one-out
cross-validation in later models.
}

