% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{okrapd}
\alias{okrapd}
\title{Example data for the documentations.}
\format{A data.table object with 912 rows and 14 variables}
\usage{
data(okrapd)
}
\description{
The data containing global okra production, area harvested and
yield from the year 1995 to 2013.
}
\keyword{datasets}

