% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paZeroEquality}
\alias{paZeroEquality}
\title{Function to check the consistency between area harvested and
production.}
\usage{
paZeroEquality(production, areaHarvested)
}
\arguments{
\item{production}{Production value}

\item{areaHarvested}{Area harvested value}
}
\description{
The function checks whether zeroes occur concordantly.
}

