% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotEnsembleOld.R
\name{plotEnsembleOld}
\alias{plotEnsembleOld}
\title{Plot Ensemble and Model Fits (Old)}
\usage{
plotEnsembleOld(data, modelFits, modelWeights, ensemble)
}
\arguments{
\item{data}{The data.table containing the data being imputed.}

\item{modelFits}{A list of length equal to the number of models.  Each
element of the list should be a numeric vector of length nrow(data). This
object is usually created by computeEnsembleFit.}

\item{modelWeights}{A matrix of dimension nrow(data) x length(modelFits).
Each element corresponds to the weight for model/column j and for
observation/row i.  Note: modelFits and modelWeights need not have the same
ordering, but they should have the same names as this is how they are
matched.}

\item{ensemble}{A numeric vector containing either the original data (if it
was not missing) or the imputed value (if the original data was missing).}

\item{value}{The column name of data containing the variable being imputed.}

\item{byKey}{The column name of data containing the grouping variable.}

\item{yearValue}{The column name of data containing the time variable.}
}
\value{
No value is returned, but a plot is generated.
}
\description{
This function plots each of the individual models in the ensemble as well
as the initial data and the final ensemble.  This function is not meant to
be called directly by the user but instead is a helper function called by
ensembleImpute (if the argument plotImputation = TRUE).  This function
differs from plotEnsemble in that it uses the original plotting code (base
graphics instead of ggplot2).
}

