% Generated by roxygen2 (4.0.1): do not edit by hand
\name{processProductionDomain}
\alias{processProductionDomain}
\title{This is a wrapper for all the data manipulation step before the
preparation of the imputation.}
\usage{
processProductionDomain(data, productionValue, areaHarvestedValue, yieldValue,
  yearValue, productionObservationFlag, areaHarvestedObservationFlag,
  yieldObservationFlag, productionMethodFlag, areaHarvestedMethodFlag,
  yieldMethodFlag, removePriorImputation = TRUE,
  removeConflictValues = TRUE, imputedFlag = "E", naFlag = "M",
  byKey = "areaCode")
}
\arguments{
\item{data}{The data}

\item{productionValue}{The name of the production variable.}

\item{areaHarvestedValue}{The name of the area harvested variable.}

\item{yieldValue}{The column name of the yield variable.}

\item{productionObservationFlag}{The observation flag of production.}

\item{productionMethodFlag}{The method flag of production.}

\item{areaHarvestedObservationFlag}{The observation flag of area
harvested.}

\item{areaHarvestedMethodFlag}{The method flag of area
harvested.}

\item{yieldObservationFlag}{The observation flag of yield.}

\item{yieldMethodFlag}{The method flag of yield.}

\item{removePriorImputation}{logical, whether prior imputation
should be removed.}

\item{removeConflictValues}{logical, whether conflict area
harvested value and production should be removed.}

\item{imputedFlag}{Flag value corresponding to values from prior
imputation, ignored if removePriorImputation is FALSE.}

\item{naFlag}{Flag value for missing values.}

\item{byKey}{The unique key identifier.}
}
\description{
This is a wrapper for all the data manipulation step before the
preparation of the imputation.
}

