% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reassignGlobalVariable.R
\name{reassignGlobalVariable}
\alias{reassignGlobalVariable}
\title{Reassign Global Variable}
\usage{
reassignGlobalVariable(var, value)
}
\arguments{
\item{var}{The name of the variable to be reassigned.}

\item{value}{The value to be reassigned to var.}
}
\value{
No value is returned, instead this function changes the global
variable var.
}
\description{
Parameters used in running the imputation or processing algorithms are
assigned to the global environment.  To protect these parameters from being
accidentally overwritten, we call lockBinding.  This prevents reassignment
of these variables, but occassionally these variables do need to be
reassigned.  This function allows this reassignment by temporarily
unlocking the variable, reassigning the value, and relocking it. This
function should be used with caution: values that have already been ensured
can be changed and not checked, which could be disastrous.
}

