% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove0M.R
\name{remove0M}
\alias{remove0M}
\title{Function to remove zero values which are missing.}
\usage{
remove0M(data, value, flag, naFlag = "M")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{value}{The value of the observation.}

\item{flag}{The flag of the observation.}

\item{naFlag}{The value of the flag which denotes missing value.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified.
}
\description{
Function to remove zero values which are missing.
}
\examples{
data = okrapd[areaName == "Bahamas",]
data[,.(productionValue, productionFlag)]
remove0M(data = data, value = "productionValue", flag = "productionFlag",
    naFlag = "M")
data[,.(productionValue, productionFlag)]
}

