% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeImputation}
\alias{removeImputation}
\title{Function to remove prior imputation.}
\usage{
removeImputation(data, value, flag, imputedFlag = "T", naFlag = "M")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{value}{The value of the observation.}

\item{flag}{The flag of the observation.}

\item{imputedFlag}{The value of the flag which denotes the value
was imputed.}

\item{naFlag}{The value of the flag which denotes missing value.}
}
\description{
Function to remove prior imputation.
}

