% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeNoInfo}
\alias{removeNoInfo}
\title{This function removes data with no information}
\usage{
removeNoInfo(data, value, flag, byKey)
}
\arguments{
\item{data}{The data}

\item{value}{The value of the production.}

\item{flag}{The flag/symbol of production.}

\item{byKey}{The unique keys which identify groups.}
}
\description{
The information contained within a set of index is calculated
based on whether a non-zero value exist. If the data contains only
zero and missing value then it is considered to contain no
imformation for imputation.
}

