% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeZeroConflict.R
\name{removeZeroConflict}
\alias{removeZeroConflict}
\title{The function removes conflicting zeroes between production and
area harvested.}
\usage{
removeZeroConflict(columnNames, naFlag = "M", data)
}
\arguments{
\item{columnNames}{See columnNames argument at ?imputeProductionDomain.}

\item{naFlag}{Flag value for missing values.}

\item{data}{The data table objest}
}
\description{
The function removes conflicting zeroes between production and
area harvested.
}

