% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testFlagTable.R
\name{testFlagTable}
\alias{testFlagTable}
\title{Test Flag Table}
\usage{
testFlagTable(flagTable, data, columnNames)
}
\arguments{
\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{data}{The data.table object containing the data.}

\item{columnNames}{See this argument at ?imputeProductionDomain.}
}
\value{
No value is returned.  However, an error is raised if the flagTable
fails the checks done by this function.
}
\description{
This function performs several checks to ensure that the provided flagTable
is valid.  First, it verifies that the column names are
"flagObservationStatus" and "flagObservationWeights".  Second, it verifies
that all observation flags in data are symbols in the flag table.
}

