% Generated by roxygen2 (4.0.1): do not edit by hand
\name{balanceAreaHarvested}
\alias{balanceAreaHarvested}
\title{Function to compute area harvested when new production and yield
are given.}
\usage{
balanceAreaHarvested(productionValue, productionObservationFlag,
  areaHarvestedValue, areaHarvestedObservationFlag, areaHarvestedMethodFlag,
  yieldValue, yieldObservationFlag, newMethodFlag,
  flagTable = faoswsFlagTable, data)
}
\arguments{
\item{productionValue}{The column name corresponding to production
value.}

\item{productionObservationFlag}{The column name corresponding to the
observation flag of production.}

\item{areaHarvestedValue}{The column name corresponding to area
harvested value.}

\item{areaHarvestedObservationFlag}{The column name corresponding
to the observation flag of area harvested.}

\item{yieldValue}{The columne name corresponding to yield value.}

\item{yieldObservationFlag}{The column name corresponding to the
observation flag of yield.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{data}{The data.table object containing the data.}
}
\description{
Function to compute area harvested when new production and yield
are given.
}

