% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/balanceProduction.R
\name{balanceProduction}
\alias{balanceProduction}
\title{Function to compute production when new area harvested and yield
are given.}
\usage{
balanceProduction(columnNames, newMethodFlag, flagTable = faoswsFlagTable,
  data)
}
\arguments{
\item{columnNames}{See this argument at ?imputeProductionDomain.}

\item{newMethodFlag}{The character value that should be assigned to
productionObservationFlag when it is imputed.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{data}{The data.table object containing the data.}
}
\description{
Function to compute production when new area harvested and yield
are given.
}

