% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeEnsembleFit.R
\name{computeEnsembleFit}
\alias{computeEnsembleFit}
\title{Function to compute the fits of all the component model}
\usage{
computeEnsembleFit(x, ensembleModel)
}
\arguments{
\item{x}{A numeric vector}

\item{ensembleModel}{A list of component models to be used to
build the ensemble.}
}
\description{
Function to compute the fits of all the component model
}

