% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeEnsembleFit.R
\name{computeEnsembleFit}
\alias{computeEnsembleFit}
\title{Function to compute the fits of all the component models}
\usage{
computeEnsembleFit(data, imputationParameters = NULL)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point. If
NULL, the parameters should have already been assigned (otherwise an error
will occur).}
}
\value{
Returns a list of vectors of the same length as ensembleModels. The
ith element of the list represents the fit of the ith model to data.
}
\description{
Function to compute the fits of all the component models
}

