% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeEnsembleWeight}
\alias{computeEnsembleWeight}
\title{Function to compute the weights of the ensemble models}
\usage{
computeEnsembleWeight(x, ensembleModel)
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{ensembleModel}{The list of ensemble models}

\item{restrictWeights}{Whether a maximum weight restriction should
be imposed.}

\item{maximumWeights}{The maximum weight to be imposed, must be
between [0.5, 1].}
}
\description{
Function to compute the weights of the ensemble models
}

