% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeErrorLOOCV.R
\name{computeErrorLOOCV}
\alias{computeErrorLOOCV}
\title{LOOCV Error Rate}
\usage{
computeErrorLOOCV(data, columnNames, value, flag, model, cvGroup)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{columnNames}{See the same argument at ?imputeProductionDomain.}

\item{value}{The column name of data which contains the values to be
imputed.}

\item{flag}{The column name of data which contains the flag describing the
status of value.}

\item{model}{The model fit to x, should be of class ensembleModel.}

\item{cvGroup}{A vector of the same length as nrow(data).  Entries of the
vector should be integers from 1 to the number of cross-validation groups
(typically 10).  This should be randomly assigned, and is usually created
by ensembleImpute.}
}
\description{
This function computes the error of different model fits via leave-one-out
cross-validation.  However, typically this function will be called via
computeErrorRate and not directly.
}

