% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeErrorLOOCV.R
\name{computeErrorLOOCV}
\alias{computeErrorLOOCV}
\title{LOOCV Error Rate}
\usage{
computeErrorLOOCV(x, model)
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{model}{The model fit to x.  If errorType="loocv" then this will be used
to compute the leave-one-out cross-validation errors.  If errorType="raw", this
is ignored.}
}
\description{
This function computes the error of different model fits via leave-one-out
cross-validation.  However, typically this function will be called via
computeErrorRate and not directly.
}
\note{
Currently, if a model fails in the cross-validation step, that error
is ignored when computing the mean error.  That could be improved by adding
a penalty when models fail to fit, or setting the error to some large value.
}

