% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeErrorRate.R
\name{computeErrorRate}
\alias{computeErrorRate}
\title{Function to compute the error of different model fits}
\usage{
computeErrorRate(x, fit, model = NULL, errorType = ifelse(is.null(model),
  "raw", "loocv"))
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{fit}{The fitted value from the model.}

\item{model}{The model fit to x.  If errorType="loocv" then this will be
used to compute the leave-one-out cross-validation errors.  If
errorType="raw", this is ignored.}

\item{errorType}{Specifies what type of error to compute.  Currently, "raw"
and "loocv" are implemented.  If "raw", then error is computed as the
difference between the model and observed data.  "loocv" performs
leave-one-out cross-validation to determine the predictive error of the
model.  "loocv" is more rigorous but much slower.}
}
\value{
A vector of (the absolute value of the) errors corresponding to the
non-missing entries in x.
}
\description{
Function to compute the error of different model fits
}
\note{
Currently, if a model fails in the cross-validation step, that error
is ignored when computing the mean error.  That could be improved by
adding a penalty when models fail to fit.
}

