% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeErrorRate}
\alias{computeErrorRate}
\title{Function to compute the mean squared error of different model fits}
\usage{
computeErrorRate(x, fit, type = "mse")
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{fit}{The fitted value from the model.}
}
\description{
Function to compute the mean squared error of different model fits
}

