% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeErrorRate}
\alias{computeErrorRate}
\title{Function to compute the error of different model fits}
\usage{
computeErrorRate(x, fit, model = NULL, errorType = ifelse(is.null(model),
  "mse", "loocv"))
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{fit}{The fitted value from the model.}

\item{model}{The model fit to x.  If errorType="loocv" then this will be used
to compute the leave-one-out cross-validation errors.  If errorType="mse", this
is ignored.}

\item{errorType}{Specifies what type of error to compute.  Currently, "mse"
and "loocv" are implemented.  "mse" computes the mean squared error between
the model and observed data.  "loocv" performs leave-one-out
cross-validation to determine the predictive error of the model.  "loocv"
is more rigorous but much slower.}
}
\description{
Function to compute the error of different model fits
}
\note{
Currently, if a model fails in the cross-validation step, that error
is ignored when computing the mean error.  That could be improved by adding
a penalty when models fail to fit.
}

