% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeErrorRate.R
\name{computeErrorRate}
\alias{computeErrorRate}
\title{Function to compute the error of different model fits}
\usage{
computeErrorRate(data, fit, imputationParameters)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{fit}{The fitted value from the model.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
A vector of (the absolute value of the) errors corresponding to the
non-missing entries in data's "value" column.
}
\description{
Function to compute the error of different model fits
}

