% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultLm.R
\name{defaultLm}
\alias{defaultLm}
\title{The default linear regression model for the ensemble model.}
\usage{
defaultLm(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default linear regression model for the ensemble model.
}

