% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultNaive.R
\name{defaultNaive}
\alias{defaultNaive}
\title{The default naive model for the ensemble model.}
\usage{
defaultNaive(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The naive model is simply linear interpolation with last
observation carried forward and backward.
}

