% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultSpline.R
\name{defaultSpline}
\alias{defaultSpline}
\title{The default spline model for the ensemble model.}
\usage{
defaultSpline(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default spline model for the ensemble model.
}

