% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ensembleImpute}
\alias{ensembleImpute}
\title{Function to perform ensemble imputation}
\usage{
ensembleImpute(x, restrictWeights = TRUE, maximumWeights = 0.7,
  ensembleModel = list(defaultMean, defaultLm, defaultExp, defaultLogistic,
  defaultLoess, defaultSpline, defaultArima, defaultMars, defaultNaive),
  plot = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{restrictWeights}{Whether a maximum weight restriction should
be imposed.}

\item{maximumWeights}{The maximum weight to be imposed, must be
between [0.5, 1].}

\item{ensembleModel}{A list of models to be used to build the
ensemble.}

\item{plot}{Whether the result of the ensemble should be plotted.}
}
\description{
This is an implementation of the ensemble imputation methodology
developed for the FAO production domain.
}

