% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{faoswsProductionImputation-package}
\alias{faoswsProductionImputation-package}
\title{Package to perform imputation for the FAOSTAT production
domain.}
\description{
This package provides the imputation methodology designed for the
FAO FAOSTAT production domain.
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

