% Generated by roxygen2 (4.0.1): do not edit by hand
\name{imputeProductionDomain}
\alias{imputeProductionDomain}
\title{This function imputes the whole production domain.}
\usage{
imputeProductionDomain(data, productionValue, areaHarvestedValue, yieldValue,
  yearValue, productionObservationFlag, areaHarvestedObservationFlag,
  yieldObservationFlag, productionMethodFlag, areaHarvestedMethodFlag,
  yieldMethodFlag, flagTable = faoswsFlagTable,
  removePriorImputation = TRUE, removeConflictValues = TRUE,
  imputedFlag = "E", imputationFlag = "I", newMethodFlag = "",
  naFlag = "M", maxdf = 5, byKey = "areaCode", restrictWeights = TRUE,
  maximumWeights = 0.7, ensembleModel = list(defaultMean = defaultMean,
  defaultLm = defaultLm, defaultExp = defaultExp, defaultLogistic =
  defaultLogistic, defaultLoess = defaultLoess, defaultSpline = defaultSpline,
  defaultArima = defaultArima, defaultMars = defaultMars, defaultNaive =
  defaultNaive), yieldFormula)
}
\arguments{
\item{data}{The data}

\item{productionValue}{The name of the production variable.}

\item{areaHarvestedValue}{The name of the area harvested variable.}

\item{yieldValue}{The column name of the yield variable.}

\item{productionObservationFlag}{The observation flag of production.}

\item{productionMethodFlag}{The method flag of production.}

\item{areaHarvestedObservationFlag}{The observation flag of area
harvested.}

\item{areaHarvestedMethodFlag}{The method flag of area
harvested.}

\item{yieldObservationFlag}{The observation flag of yield.}

\item{yieldMethodFlag}{The method flag of yield.}

\item{yearValue}{The column name corresponding to year}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{removePriorImputation}{logical, whether prior imputation
should be removed.}

\item{removeConflictValues}{logical, whether conflict area
harvested value and production should be removed.}

\item{imputedFlag}{Flag value corresponding to values from prior
imputation, ignored if removePriorImputation is FALSE.}

\item{imputationFlag}{Flag value for new imputation values.}

\item{naFlag}{Flag value for missing values.}

\item{maxdf}{The maximum degree of freedom for the spline.}

\item{byKey}{The unique key identifier.}

\item{restrictWeights}{Whether a maximum weight restriction should
be imposed.}

\item{maximumWeights}{The maximum weight to be imposed, must be
between [0.5, 1].}

\item{yieldFormula}{The formula to be passed to the linear mixed
model for the imputation of yield, if missing default spline model
is fitted.}

\item{ensembleModel}{A list of models to be used to build the
ensemble.}
}
\description{
The function will impute production, area harvested and yield at
the same time.
}
\details{
Transformation in the yield formula is not allowed and will not be
taken into account.
}

