% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeYield.R
\name{imputeYield}
\alias{imputeYield}
\title{Function to impute yield}
\usage{
imputeYield(columnNames, imputationFlag, newMethodFlag, maxdf = 5,
  flagTable = faoswsFlagTable, data, weights = NULL, yieldFormula)
}
\arguments{
\item{columnNames}{See this argument at ?imputeProductionDomain.}

\item{imputationFlag}{Flag value for new imputation values.}

\item{newMethodFlag}{The new method flag to be assigned to imputed
value.}

\item{maxdf}{The maximum degree of freedom for the spline.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}}

\item{data}{The data}

\item{weights}{The weights for the observation}
}
\description{
This function imputes the yield through linear mixed model
}

