% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeImputation.R
\name{removeImputation}
\alias{removeImputation}
\title{Function to remove prior imputation.}
\usage{
removeImputation(data, value, flag, imputedFlag = "T", naFlag = "M")
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{value}{The value of the observation.}

\item{flag}{The flag of the observation.}

\item{imputedFlag}{The value of the flag which denotes the value
was imputed.}

\item{naFlag}{The value of the flag which denotes missing value.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified.
}
\description{
Function to remove prior imputation.
}
\examples{
data = copy(okrapd[areaName=="Sudan",])
data[,.(yieldValue, yieldFlag)]
removeImputation( data = data, value = "yieldValue", flag = "yieldFlag",
    imputedFlag = "T", naFlag = "M" )
data[,.(yieldValue, yieldFlag)]
}

