% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeImputation.R
\name{removeImputation}
\alias{removeImputation}
\title{Function to remove prior imputation.}
\usage{
removeImputation(data, value, flag, processingParameters)
}
\arguments{
\item{data}{The data.table object containing the values.}

\item{value}{The value of the observation.}

\item{flag}{The flag of the observation.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified.
}
\description{
Function to remove prior imputation.
}
\examples{
data = copy(okrapd[areaName=="Sudan",])
data[,.(yieldValue, yieldFlag)]
removeImputation(data = data, value = "yieldValue", flag = "yieldFlag",
    imputedFlag = "T", naFlag = "M")
data[,.(yieldValue, yieldFlag)]
}

