% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeNoInfo.R
\name{removeNoInfo}
\alias{removeNoInfo}
\title{This function removes data with no information}
\usage{
removeNoInfo(data, value, flag, byKey, environment = parent.frame(1))
}
\arguments{
\item{data}{The data}

\item{value}{The value of the production.}

\item{flag}{The flag/symbol of production.}

\item{byKey}{The unique keys which identify groups.}

\item{environment}{The functions in this package generally work by
passing the data.table by reference and modifying it in place.  In place
row deletion is currently not supported by data.table.  Thus, to make this
function consistent with the other functions in this package, environment
is passed to specify where the modified data.table should be placed.  The
default of parent.frame(1) gives the calling environment and this should be
satisfactory for most, if not all, use cases.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified.
}
\description{
The information contained within a set of index is calculated
based on whether a non-zero value exist. If the data contains only
zero and missing value then it is considered to contain no
imformation for imputation.
}

