% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeZeroConflict.R
\name{removeZeroConflict}
\alias{removeZeroConflict}
\title{The function removes conflicting zeroes between production and
area harvested.}
\usage{
removeZeroConflict(data, processingParameters)
}
\arguments{
\item{data}{The data table objest}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}
}
\value{
No value is returned.  However, the object "data" which was passed
to this function is modified.
}
\description{
The function removes conflicting zeroes between production and
area harvested.
}
\examples{
data = okrapd[1:10,]
data[,areaHarvestedValue := c(rep(0, 5), rep(100, 5))]
data[,productionValue := c(0, 0, 100, 100, 100, 0, 0, 100, 100, 100)]
removeZeroConflict(columnNames = defaultColumnNames(), data = data)
data[,.(areaHarvestedValue, areaHarvestedFlag, productionValue,
    productionFlag, yieldValue, yieldFlag)]
}

