% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeZeroConflict}
\alias{removeZeroConflict}
\title{The function removes conflicting zeroes between production and
area harvested.}
\usage{
removeZeroConflict(productionValue, areaHarvestedValue, yieldValue,
  productionObservationFlag, areaHarvestedObservationFlag, yieldObservationFlag,
  naFlag = "M", data)
}
\arguments{
\item{productionValue}{The name of the production variable.}

\item{areaHarvestedValue}{The name of the area harvested variable.}

\item{yieldValue}{The name of the yield variable.}

\item{productionObservationFlag}{The observation flag of production.}

\item{areaHarvestedObservationFlag}{The observation flag of area
harvested.}

\item{yieldObservationFlag}{The observation flag of yield.}

\item{naFlag}{Flag value for missing values.}

\item{data}{The data table objest}
}
\description{
The function removes conflicting zeroes between production and
area harvested.
}

